+++ components/projects/index.ts
@@ -1,8 +1,0 @@
-export { CompactTaskCard } from './CompactTaskCard';
-export { ProjectCard } from './ProjectCard';
-export { TaskItem } from './TaskItem';
-export { TaskListView } from './TaskListView';
-export { DraggableTaskCard } from './DraggableTaskCard';
-export { FolderCard } from './FolderCard';
-export { FolderBreadcrumb } from './FolderBreadcrumb';
-export { MonthlyTaskScheduler } from './MonthlyTaskScheduler';